Moved `fromBandcampSchema` method from the class
