+++ src/app/release.js
@@ -89,36 +89,6 @@ export class Release {
     return this.releaseItem.title;
   }
 
-  /**
-   * @param {Object} schema
-   * @returns {Release}
-   */
-  static fromBandcampSchema(schema) {
-    const artist = schema.byArtist.name;
-    const title = schema.name;
-    const label = schema.publisher.name;
-    const datePublished = new Date(schema.datePublished);
-    const tracks = schema.track.itemListElement.map(track => new Track(
-      track.position,
-      track.item.name,
-      TrackDuration.fromDuration(track.item.duration)
-    ));
-    const url = schema.mainEntityOfPage;
-    const image = schema.image;
-    const keywords = schema.keywords;
-
-    return new Release(
-      artist,
-      title,
-      label,
-      datePublished,
-      tracks,
-      url,
-      image,
-      keywords
-    );
-  }
-
   toObject() {
     return {
       uuid: this.releaseItem.uuid,
